llm_stability/requirements.txt
# LLM Stability experiment - Python 3.11
PyYAML>=6.0
requests>=2.28
numpy>=1.24
scipy>=1.10
tqdm>=4.65
matplotlib>=3.7
